requirements.txt
# Text-to-SQL RAG System Requirements
# Python >= 3.8 required

# Core ML/AI Libraries
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available
torch>=2.0.0  # CPU version, or torch with CUDA for GPU

# Data Processing
pandas>=2.0.0
tqdm>=4.65.0  # Progress bars

# BM25 Retrieval
rank-bm25>=0.2.2

# API and Networking
requests>=2.31.0

# Development Tools (optional but recommended)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing and Evaluation
scikit-learn>=1.3.0

# Note: Ollama must be installed separately from https://ollama.ai
# Note: Spider dataset must be downloaded separately
